Create MOOSalogScraper_GPS
Currently extracts longitude, latitude, heading, yaw, and depth values. Long/lat are the equivalent of the xy-plane for the AUV, while the yaw is the "heading" of the AUV in that same space. xyz coordinates and heading has its own framework that is not the same as the one in MOOS.